Рефакторинг модуля  map
